added controller test

+++ back/src/main/java/com/valentin_d/focusarc/dto/task/TaskCreationDto.java
@@ -7,5 +7,5 @@ import jakarta.validation.constraints.Positive;
 
 import java.time.Instant;
 
-public record TaskCreationDto(@NotNull ChapterId chapterId, @Positive short estimatedMinutes, @FutureOrPresent Instant scheduledAt) {
+public record TaskCreationDto(@NotNull ChapterId chapterId, @Positive int estimatedMinutes, @FutureOrPresent Instant scheduledAt) {
 }
